app-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-dotenv
pydantic[email]
PyJWT
python-multipart
httpx
asyncio
stripe
python-dateutil
langchain_pinecone
apscheduler
pytz